Move brotli stream, bit buffer, and font diff into their own folder.
